Makefile: add initial build logic
Wrap initial build logic into a Makefile that will converge OpenWrt,
packages and later patches (or modifications) that will be applied on top
of OpenWrt.

Starting with OpenWrt 18.06, because the routers we have at the moment are
supported in this version (as the latest version). They've been dropped in
version 19.07.

Signed-off-by: Alexandru Ardelean <ardeleanalex@gmail.com>